formating updates with tools
