+++ .editorconfig
@@ -1,0 +1,13 @@
+; top-most EditorConfig file
+root = true
+
+; Unix-style newlines
+[*]
+charset = utf-8
+end_of_line = LF
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.{php,json,neon}]
+indent_style = space
+indent_size = 4
