+++ .gitignore
@@ -1,3 +1,5 @@
 /.idea/
 /vendor/
 /composer.lock
+/.php-cs-fixer.cache
+/.php-version
