+++ composer.json
@@ -18,11 +18,18 @@
     "require": {
         "php": "^8.1",
         "codeception/codeception": "^5.0",
-        "zenstruck/foundry": "^1.36"
+        "zenstruck/foundry": "^1.36",
+        "symfony/http-kernel": "^6.4 || ^7.0"
     },
     "require-dev": {
         "codeception/module-doctrine2": "^3.0",
-        "codeception/module-symfony": "^3.1"
+        "codeception/module-symfony": "^3.1",
+        "phpstan/phpstan": "^1.10",
+        "phpstan/extension-installer": "^1.3",
+        "friendsofphp/php-cs-fixer": "^3.9",
+        "phpmd/phpmd": "^2.15",
+        "php-parallel-lint/php-parallel-lint": "^1.3",
+        "php-parallel-lint/php-console-highlighter": "^1.0"
     },
     "autoload": {
         "classmap": ["src/"]
@@ -33,6 +40,9 @@
         }
     },
     "config": {
-        "classmap-authoritative": true
+        "classmap-authoritative": true,
+        "allow-plugins": {
+            "phpstan/extension-installer": true
+        }
     }
 }
