+++ phpmd.dist.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0"?>
+<ruleset name="Module Foundry"
+         xmlns="http://pmd.sf.net/ruleset/1.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://pmd.sf.net/ruleset/1.0.0
+                     http://pmd.sf.net/ruleset_xml_schema.xsd"
+         xsi:noNamespaceSchemaLocation="
+                     http://pmd.sf.net/ruleset_xml_schema.xsd">
+    <description>Ruleset for PHP Mess Detector that enforces the coding standards set forth by Salina Vortex Corporation SoftDev Team.</description>
+
+    <!-- Codeception Ruleset -->
+
+    <!-- Import entire code size rule set, modify NPath Complexity rule -->
+    <rule ref="rulesets/codesize.xml"/>
+
+    <rule ref="rulesets/cleancode.xml">
+        <exclude name="StaticAccess"/>
+    </rule>
+
+    <rule ref="rulesets/design.xml"/>
+
+    <rule ref="rulesets/naming.xml"/>
+
+    <rule ref="rulesets/unusedcode.xml">
+        <!-- PHPMD cannot recognize parameters that are enforced by an interface -->
+        <exclude name="UnusedFormalParameter"/>
+    </rule>
+</ruleset>
