+++ phpstan.dist.neon
@@ -1,0 +1,5 @@
+parameters:
+    level: 5
+    paths:
+        - src
+    excludePaths: []
