+++ src/Codeception/Module/Foundry.php
@@ -55,78 +55,8 @@ use Zenstruck\Foundry\Test\TestState;
  *             cleanup: true
  * ```
  */
-class Foundry extends Module implements DependsOnModule, RequiresPackage
+class Foundry extends AbstractFoundryConfiguration
 {
-    protected array $config = [
-        'cleanup' => false,
-        'factories' => null
-    ];
-
-    protected $dependencyMessage = <<<EOF
-        ORM module (like Doctrine2) or Framework module with ActiveRecord support is required:
-        --
-        modules:
-            enabled:
-                - Foundry:
-                    depends: Doctrine2
-        --
-    EOF;
-
-    /**
-     * @var Factory
-     */
-    public $foundry;
-
-    /**
-     * ORM module on which we depend on.
-     *
-     * @var Doctrine2
-     */
-    public $ormModule;
-
-    public function _afterSuite()
-    {
-        if ($this->getCleanupConfig() === false) {
-            return;
-        }
-
-        $this->debugSection('Foundry', 'Resetting database schema.');
-        DatabaseResetter::resetSchema($this->getSymfonyKernel());
-    }
-
-    public function _beforeSuite($settings = [])
-    {
-        $this->debugSection('Foundry', 'Booting foundry.');
-        /** @var ContainerInterface $container */
-        $container = $this->getSymfonyContainer();
-        TestState::bootFromContainer($container);
-    }
-
-    public function _depends(): array
-    {
-        return ['Codeception\Lib\Interfaces\ORM' => $this->dependencyMessage];
-    }
-
-    public function _inject(ORM $orm)
-    {
-        $this->ormModule = $orm;
-    }
-
-    public function _requires(): array
-    {
-        return [
-            'Zenstruck\Foundry\Factory' => '"zenstruck/foundry": "^1.36"',
-        ];
-    }
-
-    public function onReconfigure($settings = []): void
-    {
-        if ($this->getCleanupConfig()) {
-            DatabaseResetter::resetSchema($this->getSymfonyKernel());
-        }
-        $this->_beforeSuite($settings);
-    }
-
     /**
      * Generates and saves a record.
      *
@@ -166,7 +96,7 @@ class Foundry extends Module implements DependsOnModule, RequiresPackage
     {
         $factoryClass = $this->getFactoryClassByEntityClass($entity);
         /** @var ModelFactory $factory */
-        $factory = new $factoryClass;
+        $factory = new $factoryClass();
         $proxies = $factory->createMany($times, $attributes);
         return $this->getEntitiesByProxies($proxies);
     }
@@ -191,7 +121,7 @@ class Foundry extends Module implements DependsOnModule, RequiresPackage
     {
         $factoryClass = $this->getFactoryClassByEntityClass($entity);
         /** @var ModelFactory $factory */
-        $factory = new $factoryClass;
+        $factory = new $factoryClass();
         $memoryFactory = $factory->withoutPersisting();
         return $memoryFactory->create($attributes)->object();
     }
@@ -206,66 +136,9 @@ class Foundry extends Module implements DependsOnModule, RequiresPackage
     {
         $factoryClass = $this->getFactoryClassByEntityClass($entity);
         /** @var ModelFactory $factory */
-        $factory = new $factoryClass;
+        $factory = new $factoryClass();
         $memoryFactory = $factory->withoutPersisting();
         $proxies = $memoryFactory->createMany($times, $attributes);
         return $this->getEntitiesByProxies($proxies);
-    }
-
-    protected function getCleanupConfig(): bool
-    {
-        return $this->config['cleanup'] && $this->ormModule->_getConfig('cleanup');
-    }
-
-    /**
-     * @param Proxy[] $proxies
-     * @return object[]
-     */
-    protected function getEntitiesByProxies(array $proxies): array
-    {
-        $entities = [];
-        foreach ($proxies as $proxy) {
-            $entities[] = $proxy->object();
-        }
-        return $entities;
-    }
-
-    protected function getFactoryClassByEntityClass(string $entity): ?string
-    {
-        foreach ($this->config['factories'] as $factory) {
-            try {
-                $modelFactory = new ReflectionClass($factory);
-                $getClassMethod = $modelFactory->getMethod('getClass');
-                $getClassMethod->setAccessible(true);
-                $entityName = $getClassMethod->invoke(null);
-                if ($entity === $entityName) {
-                    return $factory;
-                }
-            } catch (ReflectionException $e) {
-                $this->fail($e->getMessage());
-            }
-        }
-        return null;
-    }
-
-    protected function getModuleSymfony(): ?Symfony
-    {
-        try {
-            /** @var Symfony $symfonyModule */
-            $symfonyModule =  $this->getModule('Symfony');
-            return $symfonyModule;
-        } catch (Exception $exception) {
-        }
-        return null;
-    }
-
-    protected function getSymfonyContainer(): SymfonyContainerInterface
-    {
-        return $this->getModuleSymfony()->_getContainer();
-    }
-
-    protected function getSymfonyKernel(): KernelInterface
-    {
-        return $this->getModuleSymfony()->kernel;
     }
 }
